Move InstallMethod to multirust::install
This encompasses installation in mulirust-specific ways,
whereas multirust-dist is all about installation from
the dist server.
