+++ src/multirust-dist/src/install.rs
@@ -1,9 +1,3 @@
-use utils;
-use errors::*;
-use temp;
-use dist;
-use component::{Components, TarGzPackage, Transaction, Package};
-
 use std::path::{Path, PathBuf};
 
 const REL_MANIFEST_DIR: &'static str = "lib/rustlib";
@@ -12,78 +6,6 @@ const REL_MANIFEST_DIR: &'static str = "lib/rustlib";
 pub struct InstallPrefix {
     path: PathBuf,
 }
-#[derive(Debug)]
-pub enum InstallMethod<'a> {
-    Copy(&'a Path),
-    Link(&'a Path),
-    Installer(&'a Path, &'a temp::Cfg),
-    Dist(&'a str, Option<&'a Path>, dist::DownloadCfg<'a>),
-}
-
-impl<'a> InstallMethod<'a> {
-    pub fn run(self, prefix: &InstallPrefix, notify_handler: NotifyHandler) -> Result<()> {
-        if prefix.is_installed_here() {
-            // Don't uninstall first for Dist method
-            match self {
-                InstallMethod::Dist(_, _, _) |
-                InstallMethod::Installer(_, _) => {}
-                _ => {
-                    try!(prefix.uninstall(notify_handler));
-                }
-            }
-        }
-
-        match self {
-            InstallMethod::Copy(src) => {
-                try!(utils::copy_dir(src, &prefix.path, ntfy!(&notify_handler)));
-                Ok(())
-            }
-            InstallMethod::Link(src) => {
-                try!(utils::symlink_dir(src, &prefix.path, ntfy!(&notify_handler)));
-                Ok(())
-            }
-            InstallMethod::Installer(src, temp_cfg) => {
-                InstallMethod::tar_gz(src, prefix, &temp_cfg, notify_handler)
-            }
-            InstallMethod::Dist(toolchain, update_hash, dl_cfg) => {
-                let maybe_new_hash =
-                    try!(dist::update_from_dist(
-                        dl_cfg,
-                        update_hash,
-                        toolchain,
-                        prefix,
-                        &[], &[]));
-
-                if let Some(hash) = maybe_new_hash {
-                    if let Some(hash_file) = update_hash {
-                        try!(utils::write_file("update hash", hash_file, &hash));
-                    }
-                }
-
-                Ok(())
-            }
-        }
-    }
-
-    fn tar_gz(src: &Path, prefix: &InstallPrefix, temp_cfg: &temp::Cfg,
-              notify_handler: NotifyHandler) -> Result<()> {
-        notify_handler.call(Notification::Extracting(src, prefix.path()));
-
-        let installation = try!(Components::open(prefix.clone()));
-        let package = try!(TarGzPackage::new_file(src, temp_cfg));
-
-        let mut tx = Transaction::new(prefix.clone(), temp_cfg, notify_handler);
-
-        for component in package.components() {
-            tx = try!(package.install(&installation, &component, None, tx));
-        }
-
-        tx.commit();
-
-        Ok(())
-    }
-}
-
 impl InstallPrefix {
     pub fn from(path: PathBuf) -> Self {
         InstallPrefix {
@@ -110,18 +32,5 @@ impl InstallPrefix {
         let mut path = PathBuf::from(REL_MANIFEST_DIR);
         path.push(name);
         path
-    }
-    fn is_installed_here(&self) -> bool {
-        utils::is_directory(&self.path)
-    }
-    pub fn uninstall(&self, notify_handler: NotifyHandler) -> Result<()> {
-        if self.is_installed_here() {
-            Ok(try!(utils::remove_dir("install", &self.path, ntfy!(&notify_handler))))
-        } else {
-            Err(Error::NotInstalledHere)
-        }
-    }
-    pub fn install(&self, method: InstallMethod, notify_handler: NotifyHandler) -> Result<()> {
-        method.run(self, notify_handler)
     }
 }
